refactor(browser): split widget input handling
